Move spacesStorage down into subdirectory
