+++ components/experimental/spaces/src/spacesStorage/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright (c) Microsoft Corporation. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+export * from "./spacesStorage";
+export * from "./spacesStorageView";
